pyproject.toml
[project]
name = "server-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "numpy>=2.4.0",
    "onnxruntime>=1.23.2",
    "opencv-python>=4.11.0.86",
    "pillow>=12.1.0",
    "ultralytics>=8.3.248",
]
